+++ app.py
@@ -8,6 +8,7 @@ from core.logs.log_handler import log_handler
 from core.mail.mail import init_mail
 from core.database.database import init_db
 from core.routes import callback_namespace, debug_namespace, trial_network_namespace, user_namespace, verification_token_namespace
+from core.utils.file_handler import load_toml
 
 app = Flask(__name__)
 CORS(app)
@@ -17,11 +18,12 @@ app.config.from_object(FlaskConf)
 
 init_db(app)
 init_mail(app)
+__version__ = load_toml("pyproject.toml", "rt", "utf-8")["tool"]["poetry"]["version"]
 
 api = Api(
     app=app,
     title=TnlcmSettings.TITLE,
-    version=TnlcmSettings.VERSION,
+    version=__version__,
     description=TnlcmSettings.DESCRIPTION,
     # doc=TnlcmSettings.DOC
 )
@@ -32,4 +34,4 @@ api.add_namespace(trial_network_namespace, path="/tnlcm/trial-network")
 api.add_namespace(user_namespace, path="/tnlcm/user")
 api.add_namespace(verification_token_namespace, path="/tnlcm/verification-token")
 
-log_handler.info(f"Start Trial Network Lifecycle Manager (TNLCM) {TnlcmSettings.VERSION} on http://0.0.0.0:{TnlcmSettings.TNLCM_PORT}")
+log_handler.info(f"Start Trial Network Lifecycle Manager (TNLCM) {__version__} on http://0.0.0.0:{TnlcmSettings.TNLCM_PORT}")